+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/standard/BaselineStandardService.java
@@ -9,7 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import gov.healthit.chpl.certificationCriteria.CertificationCriterion;
-import gov.healthit.chpl.domain.CertifiedProductSearchDetails;
 import gov.healthit.chpl.exception.EntityRetrievalException;
 import gov.healthit.chpl.util.DateUtil;
 import lombok.extern.log4j.Log4j2;
@@ -27,15 +26,36 @@ public class BaselineStandardService {
         this.standardDao = standardDao;
     }
 
-    public List<Standard> getBaselineStandardsForCriteriaAndListing(CertifiedProductSearchDetails listing, CertificationCriterion criterion, LocalDate standardCheckDate) {
+    public List<Standard> getBaselineStandards(CertificationCriterion criterion,
+            LocalDate standardCheckDateRangeStart, LocalDate standardCheckDateRangeEnd) {
         try {
-            List<StandardCriteriaMap> maps = standardDao.getAllStandardCriteriaMap();
-            Map<String, List<Standard>> standardGroups = standardGroupService.getGroupedStandardsForCriteria(criterion, standardCheckDate);
+            List<StandardCriteriaMap> stdCriteriaMaps = standardDao.getAllStandardCriteriaMap();
+            Map<String, List<Standard>> standardGroups = standardGroupService.getGroupedStandardsForCriteria(criterion, standardCheckDateRangeStart, standardCheckDateRangeEnd);
 
-            maps.removeIf(map -> !map.getCriterion().getId().equals(criterion.getId()));
-            return maps.stream()
+            stdCriteriaMaps.removeIf(map -> !map.getCriterion().getId().equals(criterion.getId()));
+            return stdCriteriaMaps.stream()
+                    .filter(stdCriteriaMap -> !isStandardInAGroup(standardGroups, stdCriteriaMap.getStandard())
+                            && DateUtil.isDateBetweenInclusive(
+                                    Pair.of(stdCriteriaMap.getStandard().getRequiredDay(), stdCriteriaMap.getStandard().getEndDay()),
+                                    standardCheckDateRangeEnd))
+                    .map(map -> map.getStandard())
+                    .toList();
+        } catch (EntityRetrievalException e) {
+            LOGGER.info("Error retrieving Standards for Criterion");
+            throw new RuntimeException(e);
+        }
+    }
+
+    public List<Standard> getActiveBaselineStandardsForCriterion(CertificationCriterion criterion,
+            LocalDate standardCheckDateRangeStart, LocalDate standardCheckDateRangeEnd) {
+        try {
+            List<StandardCriteriaMap> standardCriteriaMaps = standardDao.getAllStandardCriteriaMap();
+            Map<String, List<Standard>> standardGroups = standardGroupService.getGroupedStandardsForCriteria(criterion, standardCheckDateRangeStart, standardCheckDateRangeEnd);
+
+            standardCriteriaMaps.removeIf(map -> !map.getCriterion().getId().equals(criterion.getId()));
+            return standardCriteriaMaps.stream()
                     .filter(map -> !isStandardInAGroup(standardGroups, map.getStandard())
-                            && DateUtil.isDateBetweenInclusive(Pair.of(map.getStandard().getRequiredDay(), map.getStandard().getEndDay()), standardCheckDate))
+                            && DateUtil.isDateBetweenInclusive(Pair.of(map.getStandard().getStartDay(), map.getStandard().getEndDay()), standardCheckDateRangeStart))
                     .map(map -> map.getStandard())
                     .toList();
         } catch (EntityRetrievalException e) {